Refactoring of WeakCommand and adding WeakDelegate.
This PR refactors the WeakCommand class to add a type parameter to it
so the command handlers can be typesafe.

It also adds a new class, WeakDelegate, which is the class that
encapsulates the "trick" of having a delegate that doesn't keep the
target object alive by storing a weak reference to it, instead of what
the built-in delegate type does which is to store a strong reference
to the target.