Support only the .NET Core CLR.
This change removes the automatic support for the Mono runtime from
the extension, if the project can't run on Core CLR then it will show
an error message to the user to that respect. In that error message we
can include a link to instructions on how to either change the project
to support Core CLR or how to use the Mono image manually to deploy to
AppEngine.
This change also uses the new `dotnet` image as the base for the
generated `Dockerfile` for the deployment.