Adding support for the ActivityLog.
This commit adds support for the ActivityLog that Visual Studio
provides for extensions to write to to help with diagnostics of issues
that happen on customer's machines. This support is implemented via a
new class `ActivityLogUtils` which acts as a destination where the
extension can log events as they happen.

This commit also renames the class `AppEngineOutputWindow` to
`GcpOutputWindow` which is more in line with the rest of the extension
as well as cleaning up the class, which was implemented as an
unnecessary singleton.

There are also some minor renames of methods, logging is an area that
needs a bit of cleanup in the extension.