Adding component missing error dialog.
This commit adds a new dialog to display a more actionable error
message when missing gcloud SDK components are found, it gives the
user the exact command to run to fix their installation.

The next step once this is in will be to actually perform this
operation for the user.

Another future improvement will be to install gcloud SDK per user for
the user if we detect that it is not currently installed.