Adding support for reading auth codes from the browser.
This commit adds support for launching a browser to get the oauth
credentials and adds a simple listener (using HttpListener) for the
result of the oauth login sent from the server.

This commit supports proper cancellation of the operation and supports
redirecting to different pages depending on success or error of the
login operation.