Store the current project ID and account in the .suo file.
This change stores the current project ID and account name in the .suo
file for the project that the user has opened. The last used
project/account name pair are also persisted. When Visual Studio is
opened, before a solution is opened, these crendentials will be used
to initialize the Cloud Explorer.

This change contains a big refactoring, the current project and
account are now stored in a centralized location. Any other module can
fetch these from the centralized location. Changes to the credentials
will raise events so any sub-module that requires them can be
refreshed.