Adding password reset functionality.
This change adds password reset functionality for Windows VMs. GCloud
is used to perform the password reset as the process is quite complex
and a reimplementation in C# was not necessary.

The change adds also UI to ask the user to confirm the reset password,
and to show the new password to the user in a safe manner.