Implementing propert naming of items in Properties window.
This change implements the necessary interfaces to the Properties
window shows the proper item names for the selected items in the Cloud
Explorer.

The interface implemented is the ICustomTypeDescriptor interface and
through it the item can provice the class name and the component
name. The Properties window in Visual Studio will then show this in
the Object list. The component name will be shown in bold, it
represents the "name" of the item. The class name will be shown in
normal text, it represents the properties being shown.