Implementation of firewall manager.
This change implements UI and code behind to allow the user to enable,
and disable, fireall ports for a given GCE instance. The way it is
implemented matches what click 2 deploy does with creating firewall
rules automatically and associating them with tags set on the
instances being affected.

The code will not garbage collect the firewall rules that are no
longer necessary.