Common appearance for all dialogs.
This change adds a common base class for all dialogs as well as a
common control that all dialogs can use to have the same general
appearance. This centralizes the appearance and common settings for
all dialogs in a single class.

All dialogs have been updated to use this new base class and common
control.