Updating icons for the Cloud Explorer.
This change updates most all of the icons for the Cloud Explorer, with
icons that have transparency. These icons come from the recommended
set from Visual Studio.

This change also adds a warning icon to the Cloud Explorer, that way
messages such as "No Buckets Found" will use the warning icon instead
of nothing, which does not look good.