Adding color theme support for the Cloud Explorer.
This change adds full color theme support to the cloud explorer, which
allows the colors to be dynamically chosen depending on the currently
selected theme in VS.

There's also a new class `ThemeManager` which is able to detect the
current theme. This class can be used when different images, or more
complicated choice of colors have to be done depending on the current
them.