Use stored credentials to launch Terminal Server session (#171)
This PR adds support for using stored Windows VM credentials to launch a Terminal Server session. This is accomplished by creating a new .rdp file, a Terminal Server connection file, which contains the address of the machine to connect to, the user name and the encrypted password.

The password is encrypted by using the System.Security.Cryptography.ProtectedData class. The Terminal Server client (mstsc.exe) expects the password in the .rdp file encrypted in this manner.

With this PR the user will no longer need to copy the password from the show password dialog and pasting it in the Terminal Server window.