Fixing #869. We were trying to iterate over projects that do not support VS Automation (including unloaded). (#891)
* Fixing #869. We were trying to iterate over projects that do not support VS Automation (including unloaded). This throws an exception and since the commands have dynamic visibility set they stayed grayed out.
* To fix #869 had to change slightly what projects we consider valid and check for their validity before they were scanned for the contextual menu.

* Refactoring some to semi-merge ProjectHelper and IParsedProject since those are two different wrappers on top of the Automation Model Project interface.

*Adding some unit tests.

Some more improvements can be made, but those were too much for this PR.

Properly merge ProjectHelper and IParsedProject, most of their properties are the same. That would need though to add unit tests to clients of these class and interface before refactoring.
Refactor the code on ProjectHelper to be able to inject System.IO.Path and redo the tests so we don't have that dependency on ProjectHelper tests (this is not a file system dependency, it's just a path manipulation dependency).